Rename http module to http_client and drop Cargo.toml comments
Avoid shadowing the http crate by renaming the platform HTTP module. Remove
redundant manifest comments that explained workarounds for the old name.
