+++ src/platform/rust/src/lib.rs
@@ -1,3 +1,3 @@
 mod android;
-mod http;
+mod http_client;
 mod image;
